pyproject.toml
[project]
name = "leetcode-gamify"
version = "0.1.0"
description = ""
authors = [
    {name = "Puneet"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-mistralai (>=0.2.10,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "discord-webhook (>=1.4.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
